li.zhen: add bag sell

+++ sql/player_equip.sql
@@ -2,7 +2,7 @@ USE game_world_1001;
 
 DROP TABLE IF EXISTS player_equip;
 CREATE TABLE player_equip (
-  id          int(20) unsigned NOT NULL, # serial number
+  id          bigint(20) unsigned NOT NULL, # serial number
   player_id   int(10) unsigned NOT NULL, # player_id in player_data
   equip_id    int(10) unsigned NOT NULL, # equip_id in equip_data
   level       int(10) unsigned NOT NULL,
